restructure package and modify imports in scripts

+++ lstmcpipe/onsite_mc_r0_to_dl1.py
@@ -15,7 +15,7 @@ import random
 import argparse
 import calendar
 import lstchain
-from lstmcpipe.data_management import (
+from .io.data_management import (
     check_data_path,
     get_input_filelist,
     check_and_make_dir,
@@ -296,7 +296,7 @@ def main(input_dir, config_file=None, train_test_ratio=0.5, random_seed=42, n_r0
                 if particle == 'proton':
                     queue = 'long'
                 else:
-                    queue = 'long'  # TODO change to short after prod5 check
+                    queue = 'short'
 
                 cmd = f'sbatch --parsable -p {queue} -J {job_name[particle]} ' \
                       f'-e {jobe} -o {jobo} {base_cmd} {os.path.join(dir_lists, file)}'
